Refactors layout & adds outlet to layout to change page
