+++ app/components/layout/navLinks.js
@@ -1,0 +1,7 @@
+export const navLinks = [
+  {
+    label: 'Overview',
+    to: '/',
+  },
+  { label: 'Transaction History', to: '/transaction-history' },
+]
