+++ app/routes/transaction-history.jsx
@@ -1,0 +1,3 @@
+export default function TransactionHistory() {
+  return 'transaction history'
+}
