+++ app/styles/root.css
@@ -108,6 +108,8 @@ body {
 
 .main {
   width: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 .top-bar {
