+++ styles/pages/_index.scss
@@ -69,6 +69,8 @@
 
 .main {
   width: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 .top-bar {
